refactor: 영수증 관련 정보들 POS기에 저장
- POS 클래스 구현하여 Cashier 책임 분리
- 영수증에 출력할 정보들을 POS기에 저장하도록 함
